agents/requirements.txt
# ============================================================================
# A2A + AG-UI Starter - Python Agent Dependencies (Updated for Bedrock)
# ============================================================================

# ============================================================================
# OpenTelemetry Version Management (Google ADK 호환성)
# ============================================================================
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0

# ============================================================================
# AG-UI Protocol Implementation
# ============================================================================
ag-ui-adk>=0.0.1

# ============================================================================
# A2A Protocol Implementation
# ============================================================================
a2a>=0.1.0
a2a-sdk[http-server]

# ============================================================================
# Google ADK (Agent Development Kit) - For AG-UI Compatibility
# ============================================================================
google-adk>=0.1.0

# ============================================================================
# LangGraph Framework (Research Agent)
# ============================================================================
langgraph>=0.2.0
langchain>=0.3.0
langchain-aws>=0.1.0

# ============================================================================
# AWS Services
# ============================================================================
boto3>=1.34.0
botocore>=1.34.0

# ============================================================================
# Pydantic AI Framework
# ============================================================================
pydantic-ai>=0.0.13

# ============================================================================
# Web Server
# ============================================================================
fastapi>=0.115.0
uvicorn>=0.30.0

# ============================================================================
# Data Validation & Serialization
# ============================================================================
pydantic>=2.5.0
typing-extensions>=4.8.0

# ============================================================================
# Utilities
# ============================================================================
python-dotenv>=1.0.0

# ============================================================================
# End of File
# ============================================================================